% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonempty.R
\name{nonempty}
\alias{nonempty}
\title{Create a new nonempty data structure}
\usage{
nonempty(x)
}
\arguments{
\item{x}{non-empty structure}
}
\value{
an object of class `nonempty`
}
\description{
Create a new nonempty data structure
}
