% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonempty.R
\name{Ops,nonempty,nonempty-method}
\alias{Ops,nonempty,nonempty-method}
\title{Combine nonempty with nonempty via Ops}
\usage{
\S4method{Ops}{nonempty,nonempty}(e1, e2)
}
\description{
Combine nonempty with nonempty via Ops
}
