% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonempty.R
\name{substr,nonempty,numeric,numeric-method}
\alias{substr,nonempty,numeric,numeric-method}
\title{Extract substring from nonempty}
\usage{
\S4method{substr}{nonempty,numeric,numeric}(x, start, stop)
}
\description{
Extract substring from nonempty
}
